Separate user's and root's commands
