+++ src/main.rs
@@ -1,12 +1,15 @@
-use roxy::task::{ExecResult, Task, ERR_INVALID_COMMAND};
+mod root;
+
+use root::task::{ExecResult, Task, ERR_INVALID_COMMAND};
+use roxy::common::{self, Node, NodeRequest};
 use std::{
     io::{stdin, stdout},
     process,
 };
 
 fn main() {
-    let task: Task = match serde_json::from_reader(stdin()) {
-        Ok(task) => task,
+    let nr: NodeRequest = match serde_json::from_reader(stdin()) {
+        Ok(nr) => nr,
         Err(err) => {
             log::error!("Command Error: {}", err);
             if let Err(err) =
@@ -16,6 +19,20 @@ fn main() {
             }
             process::exit(1);
         }
+    };
+
+    let arg = base64::encode(&nr.arg);
+    let task = match nr.kind {
+        Node::Hostname(cmd) => Task::Hostname { cmd, arg },
+        Node::Interface(cmd) => Task::Interface { cmd, arg },
+        Node::Ntp(cmd) => Task::Ntp { cmd, arg },
+        Node::PowerOff => Task::PowerOff(arg),
+        Node::Reboot => Task::Reboot(arg),
+        Node::Service(cmd) => Task::Service { cmd, arg },
+        Node::Sshd(cmd) => Task::Sshd { cmd, arg },
+        Node::Syslog(cmd) => Task::Syslog { cmd, arg },
+        Node::Ufw(cmd) => Task::Ufw { cmd, arg },
+        Node::Version(cmd) => Task::Version { cmd, arg },
     };
 
     let ret = task.execute();
