+++ src/root/ntp.rs
@@ -1,4 +1,4 @@
-use crate::{run_command, run_command_output};
+use super::{run_command, run_command_output};
 use anyhow::Result;
 use regex::Regex;
 use std::{
