+++ src/root/ufw.rs
@@ -1,7 +1,7 @@
 use anyhow::Result;
 use regex::Regex;
 
-use crate::{run_command, run_command_output};
+use super::{run_command, run_command_output};
 pub type AccessLists = Vec<(String, String, String, Option<String>, Option<String>)>;
 
 /// Get firewall rules. The result of command `ufw status`.
