+++ src/user.rs
@@ -1,0 +1,3 @@
+pub mod hwinfo;
+
+use super::common::run_command_output;
